# api/openapi.yaml
openapi: 3.1.0
info:
  title: Audit Ledger REST API
  description: REST adapter for Decentralized Audit & Transparency Ledger
  version: 1.2.0
servers:
  - url: http://localhost:3002/v1
    description: Versioned API (v1)
  - url: http://localhost:3002
    description: Legacy unversioned (redirects to /v1)
...
paths:
...
  /export/events.json:
    get:
      summary: Export events as JSON
      operationId: exportEventsJson
      tags: [export]
      parameters:
        - name: limit
          in: query
          schema:
            type: integer
        - name: offset
          in: query
          schema:
            type: integer
        - name: fields
          in: query
          description: Comma-separated list of fields to include
          schema:
            type: string
        - name: filter
          in: query
          schema:
            type: object
      responses:
        "200":
          description: JSON export of events
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ExportResponse"
  /export/events.csv:
    get:
      summary: Export events as CSV
      operationId: exportEventsCsv
      tags: [export]
      parameters:
        - name: limit
          in: query
          schema:
            type: integer
        - name: offset
          in: query
          schema:
            type: integer
        - name: fields
          in: query
          description: Comma-separated list of fields to include
          schema:
            type: string
        - name: filter
          in: query
          schema:
            type: object
      responses:
        "200":
          description: CSV export of events
          content:
            text/csv:
              schema:
                type: string
  /export/events/stream:
    get:
      summary: Streaming JSON export of events
      operationId: exportEventsStream
      tags: [export]
      parameters:
        - name: limit
          in: query
          schema:
            type: integer
        - name: offset
          in: query
          schema:
            type: integer
        - name: fields
          in: query
          schema:
            type: string
      responses:
        "200":
          description: Streaming JSON events
          headers:
            X-Export-Status:
              schema:
                type: string
                enum: [running, completed, failed]
...
  /events/stream:
    get:
      summary: Stream audit events with server-sent events
      operationId: streamEvents
      tags: [events]
      parameters:
        - name: Last-Event-ID
          in: header
          required: false
          schema:
            type: integer
        - $ref: "#/components/parameters/FilterType"
        - $ref: "#/components/parameters/FilterSubmitter"
        - $ref: "#/components/parameters/FilterStartTime"
        - $ref: "#/components/parameters/FilterEndTime"
      responses:
        "200":
          description: SSE event stream
          content:
            text/event-stream:
              schema:
                type: string
  /events:
    get:
      summary: List events with pagination and filtering
      description: |
        Returns a paginated list of audit events. Supports all standard
        filter query parameters (type, submitter, metadata, startTime,
        endTime, sort, order).
      parameters:
        - $ref: "#/components/parameters/Limit"
        - $ref: "#/components/parameters/Offset"
        - $ref: "#/components/parameters/FilterType"
        - $ref: "#/components/parameters/FilterSubmitter"
        - $ref: "#/components/parameters/FilterMetadata"
        - $ref: "#/components/parameters/FilterStartTime"
        - $ref: "#/components/parameters/FilterEndTime"
        - $ref: "#/components/parameters/Sort"
        - $ref: "#/components/parameters/Order"
      responses:
        "200":
          description: Paginated list of events
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EventListResponse"
        "400":
          description: Invalid filter or pagination parameters
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "429":
          $ref: "#/components/responses/TooManyRequests"
...
  /events/{index}:
    get:
      summary: Get event by sequential index
      parameters:
        - name: index
          in: path
          description: Sequential index of the event (0-based)
          required: true
          schema:
            type: integer
            minimum: 0
          description: Sequential index of the event
      responses:
        "200":
          description: Event details
          headers:
            Cache-Control:
              schema:
                type: string
              example: "max-age=300, stale-while-revalidate=600"
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    $ref: "#/components/schemas/Event"
        "400":
          description: Invalid index parameter
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Event not found
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /events/type/{type}:
    get:
      summary: Get events by type with pagination
      description: |
        Returns all events matching the given type. Supports pagination
        via limit and offset parameters.
      parameters:
        - name: type
          in: path
          description: Event type to filter by (e.g. "payment", "audit", "governance")
          required: true
          schema:
            type: string
            maxLength: 128
          description: Event type to filter by
        - $ref: "#/components/parameters/Limit"
        - $ref: "#/components/parameters/Offset"
      responses:
        "200":
          description: Paginated list of events of the given type
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EventListResponse"
        "400":
          description: Invalid type or pagination parameters
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
components:
  parameters:
    Limit:
      name: limit
      in: query
      schema:
        type: integer
        minimum: 1
        maximum: 1000
        default: 50
      description: Number of events to return per page (1–1000)
    Offset:
      name: offset
      in: query
      schema:
        type: integer
        minimum: 0
        default: 0
      description: Number of events to skip from the beginning
    FilterType:
      name: type
      in: query
      schema:
        type: string
        maxLength: 128
      description: Filter by event type (case-insensitive partial match)
    FilterSubmitter:
      name: submitter
      in: query
      schema:
        type: string
        maxLength: 128
      description: Filter by submitter address (partial match)
    FilterMetadata:
      name: metadata
      in: query
      schema:
        type: string
        maxLength: 256
      description: Filter by metadata content (partial match)
    FilterStartTime:
      name: startTime
      in: query
      schema:
        type: integer
        minimum: 0
      description: Only include events at or after this unix timestamp (seconds)
    FilterEndTime:
      name: endTime
      in: query
      schema:
        type: integer
        minimum: 0
      description: Only include events at or before this unix timestamp (seconds)
    Sort:
      name: sort
      in: query
      schema:
        type: string
        enum: [index, timestamp, event_type, submitter]
      description: Field to sort by
    Order:
      name: order
      in: query
      schema:
        type: string
        enum: [asc, desc]
        default: asc
      description: Sort order (ascending or descending)
  responses:
    TooManyRequests:
      description: Rate limit exceeded
      headers:
        Retry-After:
          schema:
            type: integer
        X-RateLimit-Limit:
          schema:
            type: integer
        X-RateLimit-Remaining:
          schema:
            type: integer
        X-RateLimit-Reset:
          schema:
            type: integer
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
  schemas:
...
    ExportResponse:
      type: object
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/Event"
        total:
          type: integer
    Event:
      type: object
      description: An immutable audit event stored on-chain
      properties:
        id:
          type: string
          description: Stable event identifier
        index:
          type: integer
          description: Sequential event index
        timestamp:
          type: integer
          description: Unix timestamp in seconds when the event was recorded
        event_type:
          type: string
          description: Category of the event (e.g. "payment", "governance")
        submitter:
          type: string
          description: Stellar address of the event submitter
        metadata:
          type: string
          description: Hex-encoded event payload
        event_hash:
          type: string
          description: SHA-256 content hash of this event
        prev_hash:
          type: string
          description: Hash of the previous event in the chain
      required:
        - id
        - index
        - timestamp
        - event_type
        - submitter
        - metadata
        - event_hash
        - prev_hash
    EventListResponse:
      type: object
      description: Paginated list of events with metadata
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/Event"
        total:
          type: integer
          description: Total number of matching events across all pages
        limit:
          type: integer
          description: Page size used
        offset:
          type: integer
          description: Offset used
      required:
        - data
        - total
        - limit
        - offset
...
    ErrorResponse:
      type: object
      description: RFC 9457 problem details response
      required: [type, title, status, detail]
      properties:
        error:
          type: object
          required:
            - code
            - message
          properties:
            code:
              type: string
              description: Stable machine-readable error code
            message:
              type: string
              description: Human-readable error message
            details:
              type: array
              description: Specific validation failures
              items:
                type: object
                required:
                  - field
                  - message
                properties:
                  field:
                    type: string
                  message:
                    type: string
      required:
        - error